Déplace variables globales de CSG et CRDS
